+++ config/dev.exs
@@ -1,0 +1,4 @@
+use Mix.Config
+
+config :assert_html,
+  log_dsl: true